+++ philo_bonus/src/init_philo.c
@@ -20,8 +20,9 @@ static void	init_philo_order(t_philoch **p_ch)
 
 	ph_ch = *p_ch;
 	i = -1;
-	while (++i < ph_ch[0].av->num)
+	while (++i < (int)ph_ch[0].av->num)
 		ph_ch[i].order = i + 1;
+
 }
 
 static void	set_philo_av_to_ph(t_philoch **p_ch, t_philoav *p_av)
@@ -31,9 +32,10 @@ static void	set_philo_av_to_ph(t_philoch **p_ch, t_philoav *p_av)
 
 	ph_ch = *p_ch;
 	i = -1;
-	while (++i < p_av->num)
+	while (++i < (int)p_av->num)
 		ph_ch[i].av = p_av;
 }
+
 
 /*
  * argv[1] - number_of_philosophers,
